requirements.txt
Flask==2.3.3
flask-cors==4.0.0
Werkzeug==2.3.7
python-dotenv==1.0.1
openai==1.30.1
